// src/types.ts
export interface MemoryMetrics {
    start: {
        heapUsed: number;
        heapTotal: number;
        external: number;
        rss: number;
    };
    end: {
        heapUsed: number;
        heapTotal: number;
        external: number;
        rss: number;
    };
    peak: {
        heapUsed: number;
        heapTotal: number;
        external: number;
        rss: number;
    };
}

export interface BenchmarkResult {
    operation: string;
    jobsProcessed: number;
    duration: number;
    throughput: number;
    averageLatency: number;
    errors: number;
    memoryMetrics: MemoryMetrics;
}